requirements.txt
beautifulsoup4==4.11.2
requests==2.28.2
python_crontab==2.7.1
python-dotenv==1.0.0
web3==5.31.3
eth-hash==0.3.1